+++ .pylintrc
@@ -150,7 +150,8 @@ disable=print-statement,
         comprehension-escape,
         line-too-long,
         fixme,
-        fixme-info
+        fixme-info,
+        duplicate-code
 
 # Enable the message, report, category or checker with the given id(s). You can
 # either give multiple identifier separated by comma (,) or put this option
