+++ craft_store/store_client.py
@@ -18,15 +18,14 @@
 
 import base64
 import json
-from typing import Any, Dict, Optional, Sequence
-from urllib.parse import urlparse
+from typing import Optional
 
-import requests
 from macaroonbakery import bakery, httpbakery
+from overrides import overrides
 from pymacaroons.serializers import json_serializer
 
 from . import endpoints, errors
-from .auth import Auth
+from .base_client import BaseClient
 from .http_client import HTTPClient
 
 
@@ -68,9 +67,10 @@ class WebBrowserWaitingInteractor(httpbakery.WebBrowserInteractor):
         )
 
 
-class StoreClient(HTTPClient):
+class StoreClient(BaseClient):
     """Encapsulates API calls for the Snap Store or Charmhub."""
 
+    @overrides
     def __init__(
         self,
         *,
@@ -80,33 +80,21 @@ class StoreClient(HTTPClient):
         user_agent: str,
         environment_auth: Optional[str] = None,
     ) -> None:
-        """Initialize the Store Client.
-
-        :param base_url: the base url of the API endpoint.
-        :param endpoints: :data:`.endpoints.CHARMHUB` or :data:`.endpoints.SNAP_STORE`.
-        :param application_name: the name application using this class, used for the keyring.
-        :param user_agent: User-Agent header to use for HTTP(s) requests.
-        :param environment_auth: environment variable to use for credentials.
-        """
-        super().__init__(user_agent=user_agent)
+        super().__init__(
+            base_url=base_url,
+            endpoints=endpoints,
+            application_name=application_name,
+            user_agent=user_agent,
+            environment_auth=environment_auth,
+        )
 
         self._bakery_client = httpbakery.Client(
             interaction_methods=[WebBrowserWaitingInteractor(user_agent=user_agent)]
         )
-        self._base_url = base_url
-        self._store_host = urlparse(base_url).netloc
-        self._endpoints = endpoints
 
-        self._auth = Auth(application_name, base_url, environment_auth=environment_auth)
-
-    def _get_macaroon(self, token_request: Dict[str, Any]) -> str:
-        token_response = super().request(
-            "POST",
-            self._base_url + self._endpoints.tokens,
-            json=token_request,
-        )
-
-        return token_response.json()["macaroon"]
+    def _get_authorization_header(self) -> str:
+        auth = self._auth.get_credentials()
+        return f"Macaroon {auth}"
 
     def _candid_discharge(self, macaroon: str) -> str:
         bakery_macaroon = bakery.Macaroon.from_dict(json.loads(macaroon))
@@ -122,7 +110,7 @@ class StoreClient(HTTPClient):
         return base64.urlsafe_b64encode(discharged_macaroons.encode()).decode("ascii")
 
     def _authorize_token(self, candid_discharged_macaroon: str) -> str:
-        token_exchange_response = super().request(
+        token_exchange_response = self.http_client.request(
             "POST",
             self._base_url + self._endpoints.tokens_exchange,
             headers={"Macaroons": candid_discharged_macaroon},
@@ -131,99 +119,6 @@ class StoreClient(HTTPClient):
 
         return token_exchange_response.json()["macaroon"]
 
-    def login(
-        self,
-        *,
-        permissions: Sequence[str],
-        description: str,
-        ttl: int,
-        packages: Optional[Sequence[endpoints.Package]] = None,
-        channels: Optional[Sequence[str]] = None,
-    ) -> str:
-        """Obtain credentials to perform authenticated requests.
-
-        Credentials are stored on the systems keyring, handled by
-        :data:`craft_store.auth.Auth`.
-
-        The list of permissions to select from can be referred to on
-        :data:`craft_store.attenuations`.
-
-        The login process requires 3 steps:
-
-        - request an initial macaroon on :attr:`.endpoints.Endpoints.tokens`.
-        - discharge that macaroon using Candid
-        - send the discharge macaroon to :attr:`.endpoints.Endpoints.tokens_exchange`
-          to obtain final authorization of the macaroon
-
-        This last macaroon is stored into the systems keyring to
-        perform authenticated requests.
-
-        :param permissions: Set of permissions to grant the login.
-        :param description: Client description to refer to from the Store.
-        :param ttl: time to live for the credential, in other words, how
-                    long until it expires, expressed in seconds.
-        :param packages: Sequence of packages to limit the credentials to.
-        :param channels: Sequence of channel names to limit the credentials to.
-        """
-        token_request = self._endpoints.get_token_request(
-            permissions=permissions,
-            description=description,
-            ttl=ttl,
-            packages=packages,
-            channels=channels,
-        )
-
-        macaroon = self._get_macaroon(token_request)
-        candid_discharged_macaroon = self._candid_discharge(macaroon)
-        store_authorized_macaroon = self._authorize_token(candid_discharged_macaroon)
-
-        # Save the authorization token.
-        self._auth.set_credentials(store_authorized_macaroon)
-
-        return self._auth.encode_credentials(store_authorized_macaroon)
-
-    def request(
-        self,
-        method: str,
-        url: str,
-        params: Dict[str, str] = None,
-        headers: Dict[str, str] = None,
-        **kwargs,
-    ) -> requests.Response:
-        """Perform an authenticated request if auth_headers are True.
-
-        :param method: HTTP method used for the request.
-        :param url: URL to request with method.
-        :param params: Query parameters to be sent along with the request.
-        :param headers: Headers to be sent along with the request.
-
-        :raises errors.StoreServerError: for error responses.
-        :raises errors.NetworkError: for lower level network issues.
-        :raises errors.NotLoggedIn: if not logged in.
-
-        :return: Response from the request.
-        """
-        if headers is None:
-            headers = {}
-
-        auth = self._auth.get_credentials()
-        headers["Authorization"] = f"Macaroon {auth}"
-
-        return super().request(
-            method,
-            url,
-            params=params,
-            headers=headers,
-            **kwargs,
-        )
-
-    def whoami(self) -> Dict[str, Any]:
-        """Return whoami json data queyring :attr:`.endpoints.Endpoints.whoami`."""
-        return self.request("GET", self._base_url + self._endpoints.whoami).json()
-
-    def logout(self) -> None:
-        """Clear credentials.
-
-        :raises errors.NotLoggedIn: if not logged in.
-        """
-        self._auth.del_credentials()
+    def _get_discharged_macaroon(self, root_macaroon: str, **kwargs) -> str:
+        candid_discharged_macaroon = self._candid_discharge(root_macaroon)
+        return self._authorize_token(candid_discharged_macaroon)
