+++ setup.cfg
@@ -31,6 +31,7 @@ install_requires =
     pydantic
     keyring
     macaroonbakery
+    overrides
     requests
 
 [options.package_data]
